Introduced NetworkingIssueVerifier to check if a failure is a potential networking issue
